move extension API into core-frontend
