+++ common/api/core-frontend.api.md
@@ -54,7 +54,6 @@ import { CodeSpec } from '@itwin/core-common';
 import { ColorDef } from '@itwin/core-common';
 import { ColorDefProps } from '@itwin/core-common';
 import { ColorIndex } from '@itwin/core-common';
-import { CommonToolbarItem } from '@itwin/appui-abstract';
 import { CompressedId64Set } from '@itwin/core-bentley';
 import { Constructor } from '@itwin/core-bentley';
 import { ContentIdProvider } from '@itwin/core-common';
@@ -293,15 +292,12 @@ import { TileHeader } from '@itwin/core-common';
 import { TileProps } from '@itwin/core-common';
 import { TileReadStatus } from '@itwin/core-common';
 import { TileVersionInfo } from '@itwin/core-common';
-import { ToolbarOrientation } from '@itwin/appui-abstract';
-import { ToolbarUsage } from '@itwin/appui-abstract';
 import { Transform } from '@itwin/core-geometry';
 import { TransformProps } from '@itwin/core-geometry';
 import { TransientIdSequence } from '@itwin/core-bentley';
 import { Tweens } from '@itwin/core-common';
 import { TxnNotifications } from '@itwin/core-common';
 import { UiAdmin } from '@itwin/appui-abstract';
-import { UiItemsProvider } from '@itwin/appui-abstract';
 import { UnitConversion } from '@itwin/core-quantity';
 import { UnitProps } from '@itwin/core-quantity';
 import { UnitsProvider } from '@itwin/core-quantity';
@@ -3113,13 +3109,6 @@ export class ExtensionHost {
     static get toolAdmin(): ToolAdmin;
     // (undocumented)
     static get viewManager(): ViewManager;
-}
-
-// @beta (undocumented)
-export class ExtensionImpl {
-    constructor(_id: string);
-    // (undocumented)
-    registerTool(tool: ToolType, onRegistered?: () => any): Promise<void>;
 }
 
 // @beta
@@ -11746,15 +11735,6 @@ export interface ToolAssistanceSection {
 
 // @public (undocumented)
 export type ToolList = ToolType[];
-
-// @beta (undocumented)
-export class ToolProvider implements UiItemsProvider {
-    constructor(tool: ToolType);
-    // (undocumented)
-    readonly id: string;
-    // (undocumented)
-    provideToolbarButtonItems(_stageId: string, stageUsage: string, toolbarUsage: ToolbarUsage, toolbarOrientation: ToolbarOrientation): CommonToolbarItem[];
-    }
 
 // @public
 export class ToolRegistry {
