+++ common/api/summary/core-extension.exports.csv
@@ -1,0 +1,3 @@
+sep=;
+Release Tag;API Item
+beta;registerTool: (tool: ToolType, onRegistered?: (() => any) | undefined) => Promise
