--- core/extension/common_types.d.ts
@@ -1,290 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-// Enums (note these are all real!)
-export {
-  BackgroundFill,
-  BackgroundMapType,
-  BatchType,
-  BisCodeSpec,
-  ChangedValueState,
-  ChangeOpCode,
-  ColorByName,
-  CommonLoggerCategory,
-  ECSqlSystemProperty,
-  ECSqlValueType,
-  ElementGeometryOpcode,
-  FeatureOverrideType,
-  FillDisplay,
-  FillFlags,
-  FontType,
-  GeoCoordStatus,
-  GeometryClass,
-  GeometryStreamFlags,
-  GeometrySummaryVerbosity,
-  GlobeMode,
-  GridOrientationType,
-  HSVConstants,
-  ImageBufferFormat,
-  ImageSourceFormat,
-  LinePixels,
-  MassPropertiesOperation,
-  MonochromeMode,
-  Npc,
-  PlanarClipMaskMode,
-  PlanarClipMaskPriority,
-  QueryRowFormat,
-  Rank,
-  RenderMode,
-  SectionType,
-  SkyBoxImageType,
-  SpatialClassifierInsideDisplay,
-  SpatialClassifierOutsideDisplay,
-  TerrainHeightOriginMode,
-  TextureMapUnits,
-  ThematicDisplayMode,
-  ThematicGradientColorScheme,
-  ThematicGradientMode,
-  TxnAction,
-  TypeOfChange,
-} from "@itwin/core-common";
-
-// Interfaces
-export type {
-  AdditionalTransformProps,
-  AffineTransformProps,
-  AmbientLightProps,
-  AnalysisStyleDisplacementProps,
-  AnalysisStyleProps,
-  AnalysisStyleThematicProps,
-  AppearanceOverrideProps,
-  AreaFillProps,
-  AuxCoordSystem2dProps,
-  AuxCoordSystem3dProps,
-  AuxCoordSystemProps,
-  BackgroundMapProps,
-  BaseReaderOptions,
-  BRepPrimitive,
-  CalloutProps,
-  CameraProps,
-  Carto2DDegreesProps,
-  CartographicProps,
-  CategoryProps,
-  CategorySelectorProps,
-  ChangedElements,
-  ChangedEntities,
-  ChangesetIdWithIndex,
-  ChangesetIndexAndId,
-  ChangesetRange,
-  ChannelRootAspectProps,
-  ClipStyleProps,
-  CodeProps,
-  ContextRealityModelProps,
-  ContextRealityModelsContainer,
-  CutStyleProps,
-  DefinitionElementProps,
-  DeletedElementGeometryChange,
-  DeprecatedBackgroundMapProps,
-  DisplayStyle3dProps,
-  DisplayStyle3dSettingsProps,
-  DisplayStyleLoadProps,
-  DisplayStyleModelAppearanceProps,
-  DisplayStyleOverridesOptions,
-  DisplayStylePlanarClipMaskProps,
-  DisplayStyleProps,
-  DisplayStyleSettingsOptions,
-  DisplayStyleSettingsProps,
-  DisplayStyleSubCategoryProps,
-  DynamicGraphicsRequest2dProps,
-  DynamicGraphicsRequest3dProps,
-  DynamicGraphicsRequestProps,
-  EcefLocationProps,
-  ElementAspectProps,
-  ElementGeometryDataEntry,
-  ElementIdsAndRangesProps,
-  ElementLoadOptions,
-  ElementLoadProps,
-  ElementProps,
-  EmphasizeElementsProps,
-  EntityIdAndClassId,
-  EntityProps,
-  EntityQueryParams,
-  EnvironmentProps,
-  ExtantElementGeometryChange,
-  ExternalSourceAspectProps,
-  FeatureAppearanceProps,
-  FeatureAppearanceProvider,
-  FeatureAppearanceSource,
-  FilePropertyProps,
-  FlatBufferGeometryStream,
-  FontMapProps,
-  FontProps,
-  FresnelSettingsProps,
-  FunctionalElementProps,
-  GeocentricTransformProps,
-  GeodeticDatumProps,
-  GeodeticEllipsoidProps,
-  GeodeticTransformProps,
-  GeographicCRSProps,
-  GeometricElement2dProps,
-  GeometricElement3dProps,
-  GeometricElementProps,
-  GeometricModel2dProps,
-  GeometricModel3dProps,
-  GeometricModelProps,
-  GeometryAppearanceProps,
-  GeometryContainmentRequestProps,
-  GeometryContainmentResponseProps,
-  GeometryPartInstanceProps,
-  GeometryPartProps,
-  GeometryPrimitive,
-  GeometryStreamEntryProps,
-  GeometryStreamHeaderProps,
-  GeometryStreamIteratorEntry,
-  GeometrySummaryOptions,
-  GeometrySummaryRequestProps,
-  GraphicsRequestProps,
-  GridFileDefinitionProps,
-  GridFileTransformProps,
-  GroundPlaneProps,
-  Helmert2DWithZOffsetProps,
-  HemisphereLightsProps,
-  HorizontalCRSExtentProps,
-  HorizontalCRSProps,
-  ImageGraphicProps,
-  ImagePrimitive,
-  InformationPartitionElementProps,
-  JsonGeometryStream,
-  LightSettingsProps,
-  LineStyleProps,
-  LocalBriefcaseProps,
-  Localization,
-  MassPropertiesRequestProps,
-  MassPropertiesResponseProps,
-  MaterialProps,
-  ModelClipGroupProps,
-  ModelGeometryChanges,
-  ModelGeometryChangesProps,
-  ModelIdAndGeometryGuid,
-  ModelLoadProps,
-  ModelProps,
-  ModelQueryParams,
-  ModelSelectorProps,
-  NavigationBindingValue,
-  NavigationValue,
-  PartReference,
-  PersistentGraphicsRequestProps,
-  PhysicalElementProps,
-  PhysicalTypeProps,
-  Placement2dProps,
-  Placement3dProps,
-  PlanarClipMaskProps,
-  PlanProjectionSettingsProps,
-  PositionalVectorTransformProps,
-  ProjectionProps,
-  QueryLimit,
-  QueryOptions,
-  QueryQuota,
-  RelatedElementProps,
-  RelationshipProps,
-  RenderMaterialAssetProps,
-  RenderMaterialProps,
-  RenderTimelineLoadProps,
-  RenderTimelineProps,
-  RepositoryLinkProps,
-  RequestNewBriefcaseProps,
-  RgbColorProps,
-  RootSubjectProps,
-  RpcActivity,
-  SectionDrawingLocationProps,
-  SectionDrawingProps,
-  SectionDrawingViewProps,
-  SessionProps,
-  SheetProps,
-  SkyBoxImageProps,
-  SkyBoxProps,
-  SkyCubeProps,
-  SolarLightProps,
-  SolarShadowSettingsProps,
-  SourceAndTarget,
-  SpatialClassifierFlagsProps,
-  SpatialClassifierProps,
-  SpatialClassifiersContainer,
-  SpatialViewDefinitionProps,
-  SubCategoryProps,
-  SubjectProps,
-  TerrainProps,
-  TextStringPrimitive,
-  TextStringProps,
-  TextureData,
-  TextureLoadProps,
-  TextureMapProps,
-  TextureProps,
-  ThematicDisplayProps,
-  ThematicDisplaySensorProps,
-  ThematicDisplaySensorSettingsProps,
-  ThematicGradientSettingsProps,
-  ThumbnailFormatProps,
-  ThumbnailProps,
-  TileVersionInfo,
-  TypeDefinitionElementProps,
-  UrlLinkProps,
-  VerticalCRSProps,
-  ViewAttachmentLabelProps,
-  ViewAttachmentProps,
-  ViewDefinition2dProps,
-  ViewDefinition3dProps,
-  ViewDefinitionProps,
-  ViewDetails3dProps,
-  ViewDetailsProps,
-  ViewFlagProps,
-  ViewQueryParams,
-  ViewStateLoadProps,
-  ViewStateProps,
-  WhiteOnWhiteReversalProps,
-  XyzRotationProps,
-} from "@itwin/core-common"
-
-// Types
-export {
-  AxisAlignedBox3d,
-  AxisAlignedBox3dProps,
-  BackgroundMapProviderName,
-  Base64EncodedString,
-  BriefcaseId,
-  ChangesetId,
-  ChangesetIndex,
-  ChangesetIndexOrId,
-  CodeScopeProps,
-  ColorDefProps,
-  DanishSystem34Region,
-  EasingFunction,
-  ElementAlignedBox2d,
-  ElementAlignedBox3d,
-  ElementGeometryChange,
-  ElementGraphicsRequestProps,
-  EntityIdAndClassIdIterable,
-  GeodeticTransformMethod,
-  GeometryStreamPrimitive,
-  GeometryStreamProps,
-  GridFileDirection,
-  GridFileFormat,
-  HemisphereEnum,
-  ImageGraphicCornersProps,
-  InterpolationFunction,
-  LocalAlignedBox3d,
-  PersistentBackgroundMapProps,
-  Placement,
-  PlacementProps,
-  Point2dProps,
-  ProjectionMethod,
-  RemoveFunction,
-  RgbFactorProps,
-  TerrainProviderName,
-  TweenCallback,
-  UnitType,
-  UpdateCallback,
-  ViewFlagOverrides,
-  ViewFlagsProperties,
-} from "@itwin/core-common"
