--- core/extension/frontend_types.d.ts
@@ -1,295 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-// Interfaces
-export type {
-  Animator,
-  BatchOptions,
-  BeButtonEventProps,
-  BeTouchEventProps,
-  BeWheelEventProps,
-  CanvasDecoration,
-  ChangeViewedModel2dOptions,
-  ComputeChordToleranceArgs,
-  CreateTextureArgs,
-  CreateTextureFromSourceArgs,
-  CustomGraphicBuilderOptions,
-  Decorator,
-  DepthPointOptions,
-  DepthRangeNpc,
-  ExtentLimits,
-  FeatureOverrideProvider,
-  FrontendSecurityOptions,
-  FuzzySearchResult,
-  GlobalAlignmentOptions,
-  GlobalLocation,
-  GlobalLocationArea,
-  GpuMemoryLimits,
-  GraphicArc,
-  GraphicArc2d,
-  GraphicBranchOptions,
-  GraphicBuilderOptions,
-  GraphicLineString,
-  GraphicLineString2d,
-  GraphicLoop,
-  GraphicPath,
-  GraphicPointString,
-  GraphicPointString2d,
-  GraphicPolyface,
-  GraphicPrimitive2d,
-  GraphicShape,
-  GraphicShape2d,
-  GraphicSolidPrimitive,
-  HitListHolder,
-  IModelIdArg,
-  MarginOptions,
-  OffScreenViewportOptions,
-  OnViewExtentsError,
-  OsmBuildingDisplayOptions,
-  ParsedKeyin,
-  ParseKeyinError,
-  ParticleCollectionBuilder,
-  ParticleCollectionBuilderParams,
-  ParticleProps,
-  PickableGraphicOptions,
-  ScreenSpaceEffectBuilder,
-  ScreenSpaceEffectBuilderParams,
-  ScreenSpaceEffectContext,
-  ScreenSpaceEffectSource,
-  SelectAddEvent,
-  SelectedViewportChangedArgs,
-  SelectRemoveEvent,
-  SelectReplaceEvent,
-  TextureCacheOwnership,
-  TextureImage,
-  TileContent,
-  TiledGraphicsProvider,
-  TileDrawArgParams,
-  TileParams,
-  TileTreeDiscloser,
-  TileTreeOwner,
-  TileTreeParams,
-  TileTreeSupplier,
-  ToolAssistanceInstruction,
-  ToolAssistanceInstructions,
-  ToolAssistanceKeyboardInfo,
-  ToolAssistanceSection,
-  ToolTipOptions,
-  Uniform,
-  UniformArrayParams,
-  UniformContext,
-  UniformParams,
-  ViewAnimationOptions,
-  ViewChangeOptions,
-  ViewClipEventHandler,
-  ViewCreator2dOptions,
-  ViewCreator3dOptions,
-  ViewportDecorator,
-  ViewportGraphicBuilderOptions,
-  ZoomToOptions,
-} from "@itwin/core-frontend";
-
-// Types
-export type {
-  CanvasDecorationList,
-  FlashSettingsOptions,
-  GpuMemoryLimit,
-  GraphicList,
-  GraphicPrimitive,
-  MarkerFillStyle,
-  MarkerImage,
-  MarkerTextAlign,
-  MarkerTextBaseline,
-  OnFlashedIdChangedEventArgs,
-  ParseKeyinResult,
-  SelectionSetEvent,
-  TextureCacheKey,
-  TextureImageSource,
-  TextureOwnership,
-  ToolList,
-  ToolType,
-} from "@itwin/core-frontend";
-
-// Enums (note these are all real!)
-export {
-  ACSDisplayOptions,
-  ACSType,
-  ActivityMessageEndReason,
-  BeButton,
-  BeModifierKeys,
-  ClipEventType,
-  ContextRotationId,
-  CoordinateLockOverrides,
-  CoordSource,
-  CoordSystem,
-  DepthPointSource,
-  EventHandled,
-  FlashMode,
-  GraphicType,
-  HitDetailType,
-  HitGeomType,
-  HitParentGeomType,
-  HitPriority,
-  HitSource,
-  InputSource,
-  KeyinParseError,
-  LocateAction,
-  LocateFilterStatus,
-  ManipulatorToolEvent,
-  MessageBoxIconType,
-  MessageBoxType,
-  MessageBoxValue,
-  OutputMessageAlert,
-  OutputMessagePriority,
-  OutputMessageType,
-  ParseAndRunResult,
-  SelectionMethod,
-  SelectionMode,
-  SelectionProcessing,
-  SelectionSetEventType,
-  SnapHeat,
-  SnapMode,
-  SnapStatus,
-  StandardViewId,
-  StartOrResume,
-  TextureTransparency,
-  TileBoundingBoxes,
-  TileGraphicType,
-  TileLoadPriority,
-  TileLoadStatus,
-  TileTreeLoadStatus,
-  TileVisibility,
-  ToolAssistanceImage,
-  ToolAssistanceInputMethod,
-  UniformType,
-  VaryingType,
-  ViewStatus,
-  ViewUndoEvent,
-} from "@itwin/core-frontend";
-
-// Abstract Classes
-export type {
-  AuxCoordSystemState,
-  DisplayStyleState,
-  GeometricModelState,
-  GraphicBuilder,
-  IModelConnection,
-  // InputCollector, @REAL
-  // InteractiveTool, @REAL
-  MarkerSet,
-  NotificationHandler,
-  // PrimitiveTool, @REAL
-  RenderClipVolume,
-  RenderGraphic,
-  RenderGraphicOwner,
-  RenderSystem,
-  Tile,
-  TileTree,
-  TileTreeReference,
-  ViewManip,
-  Viewport,
-  ViewPose,
-  ViewState,
-  ViewState2d,
-  ViewState3d,
-  // ViewTool @REAL
-} from "@itwin/core-frontend";
-
-// Classes
-export type {
-  // AccuDrawHintBuilder, @REAL
-  AccuSnap,
-  ActivityMessageDetails,
-  AuxCoordSystem2dState,
-  AuxCoordSystem3dState,
-  AuxCoordSystemSpatialState,
-  // BeButtonEvent, @REAL
-  BeButtonState,
-  BeTouchEvent,
-  BeWheelEvent,
-  // BingElevationProvider, @REAL
-  // BingLocationProvider, @REAL
-  CategorySelectorState,
-  ChangeFlags,
-  // Cluster, @REAL
-  ContextRealityModelState,
-  DecorateContext,
-  Decorations,
-  DisclosedTileTreeSet,
-  DisplayStyle2dState,
-  DisplayStyle3dState,
-  DrawingModelState,
-  DrawingViewState,
-  DynamicsContext,
-  ElementLocateManager,
-  ElementPicker,
-  ElementState,
-  // EmphasizeElements, @REAL
-  EntityState,
-  EventController,
-  FlashSettings,
-  // FrustumAnimator, @REAL
-  GeometricModel2dState,
-  GeometricModel3dState,
-  // GlobeAnimator, @REAL
-  // GraphicBranch, @REAL
-  HiliteSet,
-  HitDetail,
-  HitList,
-  IconSprites,
-  // IModelApp, @BANNED
-  IntersectDetail,
-  // IpcApp, @BANNED
-  LocateOptions,
-  // LocateResponse, @REAL
-  // MarginPercent, @REAL
-  // Marker, @REAL
-  ModelSelectorState,
-  ModelState,
-  NotificationManager,
-  // NotifyMessageDetails, @REAL
-  OffScreenViewport,
-  OrthographicViewState,
-  PhysicalModelState,
-  RenderContext,
-  Scene,
-  SceneContext,
-  ScreenViewport,
-  SectionDrawingModelState,
-  SelectionSet,
-  SheetModelState,
-  SheetViewState,
-  SnapDetail,
-  SpatialLocationModelState,
-  SpatialModelState,
-  SpatialViewState,
-  Sprite,
-  SpriteLocation,
-  TentativePoint,
-  TileAdmin,
-  TileDrawArgs,
-  TileRequest,
-  TileRequestChannel,
-  TileRequestChannels,
-  TileRequestChannelStatistics,
-  Tiles,
-  TileUsageMarker,
-  // Tool, @REAL
-  ToolAdmin,
-  // ToolAssistance, @REAL
-  ToolSettings,
-  ViewCreator2d,
-  ViewCreator3d,
-  ViewingSpace,
-  ViewManager,
-  // ViewRect, @REAL
-} from "@itwin/core-frontend";
-
-// namespaces (minus conflicts)
-export type {
-  // EditManipulator, @REAL
-  // FeatureSymbology, @REAL
-  PerModelCategoryVisibility,
-  // Pixel @REAL
-} from "@itwin/core-frontend";
