+++ core/extension/package.json
@@ -2,21 +2,26 @@
   "name": "@itwin/core-extension",
   "version": "3.1.0-dev.36",
   "description": "iTwin.js Extension API",
-  "typings": "index.d.ts",
+  "main": "lib/cjs/core-extension.js",
+  "module": "lib/esm/core-extension.js",
+  "typings": "lib/cjs/core-extension",
   "license": "MIT",
   "repository": {
     "type": "git",
     "url": "https://github.com/iTwin/itwinjs-core/tree/master/core/extension"
   },
   "scripts": {
-    "compile": "",
-    "build": "",
-    "build:ci": "",
-    "clean": "rimraf .rush/temp/package-deps*.json",
+    "build": "npm run -s build:cjs && npm run -s build:esm",
+    "build:ci": "npm run -s build && npm run -s build:esm",
+    "build:cjs": "tsc 1>&2 --outDir lib/cjs",
+    "build:esm": "tsc 1>&2 --module ES2020 --outDir lib/esm",
+    "clean": "rimraf lib .rush/temp/package-deps*.json",
+    "extract-api": "betools extract-api --entry=core-extension",
+    "lint": "eslint -f visualstudio \"./src/**/*.ts\" 1>&2",
+    "test": "",
+    "docs": "betools docs --includes=../../generated-docs/extract --json=../../generated-docs/core/core-extension/file.json --tsIndexFile=./core-extension.ts --onlyJson",
     "cover": "",
-    "docs": "",
-    "lint": "",
-    "test": ""
+    "start": "npm run build:cjs -- --watch && npm run build:esm -- --watch"
   },
   "keywords": [
     "Bentley",
@@ -29,15 +34,20 @@
     "name": "Bentley Systems, Inc.",
     "url": "http://www.bentley.com"
   },
-  "dependencies": {
-    "@itwin/core-frontend": "workspace:*",
-    "@itwin/core-common": "workspace:*"
+  "peerDependencies": {
+    "@itwin/appui-abstract": "workspace:^3.1.0-dev.36",
+    "@itwin/core-frontend": "workspace:^3.1.0-dev.36",
+    "@itwin/core-common": "workspace:^3.1.0-dev.36"
   },
   "devDependencies": {
+    "@itwin/appui-abstract": "workspace:*",
     "@itwin/build-tools": "workspace:*",
+    "@itwin/core-frontend": "workspace:*",
+    "@itwin/core-common": "workspace:*",
     "@itwin/eslint-plugin": "workspace:*",
     "eslint": "^7.11.0",
-    "rimraf": "^3.0.2"
+    "rimraf": "^3.0.2",
+    "typescript": "~4.4.0"
   },
   "eslintConfig": {
     "plugins": [
