+++ core/extension/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./node_modules/@itwin/build-tools/tsconfig-base.json",
+  "compilerOptions": {
+    "esModuleInterop": true,
+    "module": "es2020"
+  },
+  "include": ["./src/**/*.ts"]
+}
